removing example from embedded example to avoid nesting examples
